In the beginning, there was a PCO Alfred workflow
This work flow introduces the following:

- `pco` keyword to get a `PcoActionsList`
  - "Box" jumps to `PcoBoxActionsList`
  - "API Docs" browser link
  - "VictorOps" browser link
  - "Sidekiq" browser link (allows choosing app and env)
  - "Picodex" browser link
  - "Lunchbox" browser link
  - "Datadog" browser link
  - "Bugsnag" browser link
  - "Expensify" browser link
  - "MyADP" browser link
  - "ADP TotalSource" browser link
- `box` keyword to get a `PcoBoxActionsList`
  - Allow Idle (allows choosing app)
  - Background Logs
  - Check Version
  - Open local App in Browser
  - Prevent Idle (allows choosing app)
  - Restart PCO Box
  - Restart App (allows choosing app)
  - Restart Nginx
  - Start PCO Box
  - Stop PCO Box

All the code follows the basic form of a public `#to_json` method that
returns the JSON Alfred expects: `{ items: [array, of, items] }`. For
now, let's opt for a separation between the data we care about vs. the
hash keys that Alfred cares about. That way we're free to handle the
mapping between the two however we like. It might be nice to move data
out to YAML, or whatever else, if this project grows.